use 'pillow' instead of 'pil' in requirements.txt
On recent Debian/Ubuntu systems, a dynamic library bug in PIL breaks
encoding/decoding of PNG and JPEG formats:
  http://obroll.com/install-python-pil-python-image-library-on-ubuntu-11-10-oneiric/
  http://unix.stackexchange.com/questions/33664/zlib-support-for-pil-wont-work

This has dire consequences as it breaks the captcha of the
'inscription' page. Replacing PIL by its fork pillow apparently is the
recommended fix -- I personally could also use the patched Debian
package instead of installing through 'pip', but that would not be
portable.