Do not use template after validation
As we do in the login view, we cannot directly render a template after a
login because the request.user will still be anonymous user and will
make displaying user-related templates fail.

We need to redirect to any page (here homepahe) so that the request.user
will be updated and templates will render correctly.