Do not use Fabric but a Makefile
Since Fabric is not Python 3 compatible (nor won't be), using a Makefile
seems the good choice instead of maintaining both Py2 and Py3 separate
virtualenv for developers.

Meanwhile, Fabric may still be used for deployement on production server
so that's why there is a new entry in the .gitignore file.