show revealed moderated message only on click
The content of the moderated message can be shown by clicking on the
link "Message modéré par".

We reuse the implementation technique that was formerly used to
implement secret blocks: using the CSS `:target` predicate, that is
true of an element when its `id` is being referred to by the current
URL. This means that un-hiding a message jumps to the message, and
that at most one message can be un-hidden at a time.